Add IndexEntries adapter
